moved to the shaded guava31 with 1.18
