+++ flink-connector-snowflake/src/main/java/io/deltastream/flink/connector/snowflake/sink/internal/SnowflakeInternalUtils.java
@@ -17,8 +17,8 @@
 
 package io.deltastream.flink.connector.snowflake.sink.internal;
 
-import org.apache.flink.shaded.guava30.com.google.common.base.Joiner;
-import org.apache.flink.shaded.guava30.com.google.common.base.Preconditions;
+import org.apache.flink.shaded.guava31.com.google.common.base.Joiner;
+import org.apache.flink.shaded.guava31.com.google.common.base.Preconditions;
 
 import org.apache.commons.lang3.StringUtils;
